refactor style
